test: 알림 발송 및 복구 로직 통합 테스트 구현
